% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melb_trams.R
\name{melb_trams}
\alias{melb_trams}
\title{Melbourne Trams Colour Palette}
\usage{
melb_trams(n, name = NULL)
}
\arguments{
\item{n}{Number of colours that you desire in your palette.}

\item{name}{Name of any specific colours or tram numbers that you want in your palette (optional).}
}
\value{
Beautiful colour palette for any visualisation purposes.
}
\description{
Although the trams can be unreliable and the bane of our existence, the
aesthetically pleasing colour palette makes it all worth it.
}
\author{
Dionne Argyropoulos
}
